Move test-related config to rails_helper
