+++ .circleci/config.yml
@@ -100,6 +100,7 @@ jobs:
             bundle exec rspec --profile 10 \
                               --color \
                               --order random \
+                              --require rails_helper \
                               --format RspecJunitFormatter \
                               --out test-results/rspec/rspec.xml \
                               --format progress \
