+++ spec/rails_helper.rb
@@ -8,6 +8,10 @@ require File.expand_path("../../config/environment", __FILE__)
 
 require "rspec/rails"
 
+require "database_cleaner"
+require "rspec-sidekiq"
+require "vcr"
+
 ActiveRecord::Migration.maintain_test_schema!
 
 # Requires supporting ruby files with custom matchers and macros, etc,
@@ -22,7 +26,103 @@ RSpec.configure do |config|
   config.include RequestSpecHelpers, type: :controller
 end
 
-def set_current_user(user)
-  cookies.signed[:auth] =
-    { secure: true, httponly: true, value: [user.id, user.auth_token] }
+# For codeclimate test coverage.
+# Only enable if the environmental variable is set - i.e. on CI
+if ENV["COVERAGE"]
+  require "simplecov"
+  SimpleCov.start "rails"
+end
+
+VCR.configure do |config|
+  config.cassette_library_dir = "spec/vcr_cassettes"
+  config.allow_http_connections_when_no_cassette = true
+  config.hook_into :webmock
+end
+
+RSpec::Sidekiq.configure do |config|
+  config.warn_when_jobs_not_processed_by_sidekiq = false
+end
+
+# DB Cleaner metadata tags
+# ========================
+#
+# Use the following RSpec metadata tags to tweak how the database is reset
+# between tests.
+#
+# * `strategy: :transaction` (default)
+#
+# Fastest. Performs queries in a transaction and rolls back at the end of the
+# test. A sensible default. The tag can be omitted.
+#
+# * `strategy: :deletion`
+#
+# Slower, but comparable in speed to :transaction for small data sets.
+# Useful for testing `after_commit` callbacks.
+# Does not re-create tables or indexes.
+#
+# * `strategy: :truncation`
+#
+# Slowest. Fixed-time regardless of the amount of data (hence cost-effective
+# only for large or complicated data setups). Runtime grows with the number of
+# tables, indexes, complexity of the db overall.
+#
+# * `:context_state`
+#
+# To skip example-wise cleaning (e.g., to share large setup between examples).
+# Use sparingly.
+#
+# ```
+# before(:all) { ... }
+#
+# describe "test 1", :context_state do ...
+#
+# describe "test 2", :context_state do ...
+# ```
+#
+# * `:js`
+#
+# For acceptance / system tests (Capybara/Cucumber/Rails 5 system tests) with a
+# JS driver. Enables the truncation strategy, which is slowest but most stable.
+#
+class DirtyDatabaseError < RuntimeError
+  def initialize(meta)
+    super "#{meta[:full_description]}\n\t#{meta[:location]}"
+  end
+end
+
+RSpec.configure do |config|
+  config.before(:suite) do
+    DatabaseCleaner.clean_with(:deletion)
+  end
+
+  config.before(:all, :context_state) do
+    DatabaseCleaner.strategy = :truncation
+    DatabaseCleaner.start
+  end
+
+  config.before(:each) do |example|
+    next if example.metadata[:context_state]
+
+    strategy =
+      example.metadata[:strategy] ||
+      (example.metadata[:js] ? :truncation : :transaction)
+
+    DatabaseCleaner.strategy = strategy
+    DatabaseCleaner.start
+  end
+
+  config.after(:each) do |example|
+    next if example.metadata[:context_state]
+
+    DatabaseCleaner.clean
+
+    # For debugging:
+    # if ModelName.count > 0
+    #   raise DirtyDatabaseError.new(example.metadata)
+    # end
+  end
+
+  config.after(:all, :context_state) do
+    DatabaseCleaner.clean
+  end
 end
