+++ spec/support/request_spec_helpers.rb
@@ -1,3 +1,8 @@
+def set_current_user(user)
+  cookies.signed[:auth] =
+    { secure: true, httponly: true, value: [user.id, user.auth_token] }
+end
+
 RSpec.shared_context :logged_in_as_user do
   let(:user) { FactoryBot.create(:user_confirmed) }
   before { set_current_user(user) }
