--- spec/support/sidekiq.rb
@@ -1,5 +1,0 @@
-require "sidekiq/testing"
-
-RSpec::Sidekiq.configure do |config|
-  config.warn_when_jobs_not_processed_by_sidekiq = false
-end
