--- spec/support/simplecov.rb
@@ -1,6 +1,0 @@
-# For codeclimate test coverage.
-# Only enable if the environmental variable is set - i.e. on CI
-if ENV["COVERAGE"]
-  require "simplecov"
-  SimpleCov.start "rails"
-end
